faking migration information for compatibility with rails in use
I am pretty sure that these migration were likely created when
virginia was using a Rails 4 version of the app? But my memory is
poor and the Gemfile.lock at the time these were committed to the repo
says 5.0, so I wedged that in to see if it will work.